refactor: benchmarks + add baseline profile generator
