+++ depconstraints/build.gradle.kts
@@ -56,7 +56,7 @@ val hamcrest = "1.3"
 val hilt = Versions.HILT_AGP
 val junit = "4.13.2"
 val junitExt = "1.1.1"
-val lifecycle = "2.4.0-alpha01"
+val lifecycle = "2.4.1"
 val lottie = "3.0.0"
 val material = "1.4.0-beta01"
 val mockito = "3.3.1"
